foudationdb/untangle: remove spawn without lock...
... spawn without lock does not make sense. If the resource must be
lock when access it must be lock anytime it is accessed.  Previously
it was only locked when a task is poped and when another thread adds a
task, and not when the main thread adds a task. That would lead to a
memory corruption or something bad.

So now, it is slow but correct :)